refactor: api version tag and refactor structure passenger service :hammer:
