--- src/Services/Passenger/Passenger.Application/IPassengerRepository.cs
@@ -1,8 +1,0 @@
-﻿using Common.Application.Repositories;
-
-namespace Passenger.Application
-{
-    public interface IPassengerRepository : IBaseRepository<Passenger.Domain.Entities.Passenger>
-    {
-    }
-}
