+++ src/Services/Passenger/Passenger.Application/Repositories/IPassengerRepository.cs
@@ -1,0 +1,8 @@
+﻿using Common.Application.Repositories;
+
+namespace Passenger.Application.Repositories
+{
+    public interface IPassengerRepository : IBaseRepository<Domain.Entities.Passenger>
+    {
+    }
+}
